Pull request: all: upd dnsproxy, fix test, minor imp
Merge in DNS/adguard-home from upd-dnsproxy to master

Squashed commit of the following:

commit 52365328a4c653b0d3bf6d11d8f66a3bf8e0f402
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Thu Jun 10 18:49:50 2021 +0300

    all: rm debug mod replace

commit b29e3cffc5dfa58951a9d819a196e588d26dd887
Author: Ainar Garipov <A.Garipov@AdGuard.COM>
Date:   Thu Jun 10 18:46:22 2021 +0300

    all: upd dnsproxy, fix test, minor imp

+++ internal/dhcpd/options.go
@@ -130,5 +130,5 @@ func (p *dhcpOptionParser) parse(s string) (code uint8, data []byte, err error) 
 		return 0, nil, err
 	}
 
-	return uint8(code), data, nil
+	return code, data, nil
 }
